Put in latest TRD geometry based on consultation with Aachen people. Octagon structure has been rearranged somewhat: there are now 5 "primary" octagons.  Top and bottom carbon rings are separate from the top and bottom honeycomb octagons and the central octagon. The ladder dimension calculation has been checked against Aachen engineer's numbers.
Because most of the TRD will be stuffed with radiator, including
various corners at the sides, I have made the octagon which is mother
to the ladders into a radiator volume; vacuum "holes" now take the
place of some of the old radiators at the bottom.  Note that
"radiator" volumes are no longer made of radiator material.  This
scheme may be changed if necessary.

Still to be simulated: bulkheads, strips and spacers, manifolds,
notches in bottom octagon.

There's still possibly some problem with overlap with other volumes
at the bottom, but I think absolute TRD positioning is correct here.