Added better fix for gcc 4.x x>4
The problem was depending on the linking order,
the gROOT global variable would not be properly
initialized before it's first use causing an
abort. The better fix is to explicitly initialize
gROOT if it hasn't been defined before the
offending line.

Since this is a better fix, I removed
the linker flags I added earlier.