Tracker charge calibration improvement.
Resolution improves a bit, and several residual dependence are reduced.
This will simplify charge distinction, template fitting and PDF estimation.

Quick how-to: the interface is not changed.
> float Q_trk = track->GetQ(beta).
An additional correction based on rigidity for the relativistic rise can be used
> float Q_trk = track->GetQ(beta,fit_id,m_on_Z);
For a good rigidity correction you you have to assign the M/Z parameter
otherwise the algorithm will make an educated guess.
To use the old computation you have to put in the last field the version number:
> float Q_old = track->GetQ(beta,-1,0,1);