Change vector of registered G4 tracks from vector<int> to set<int>.
This removes an enormous performance bottle neck for particles with very many
tracks (say, a 4TeV proton interacting in the ECAL).

Reason is that find() is O(n) in vector, but O(log(n)) in a set.

A set is more natural to use here, anyway.

Performance study details will be provided by N. Zimmermann.