Encode process identifier into MCEventgR::Nskip and TrdMCClusterR::ProcID.
For secondary particles, Nskip now contains the full information about the
G4Process which created the particle. The format is as follows:

Nskip == (G4ProcessType << 24) | (G4ProcessSubType & 0xFFFFFF)

The same format is also used to store some information about which process
created the particle responsible for TrdMCClusters. This is needed to
differentiate between transition radiation and delta ray signals.