Fix crash on destruction related to TOF MC simulation.
Don't create THxx objects on the stack, if the AMS ROOT file is already open.
This will associate the histograms with the TFile, and once the TFile is Close()d
it will try to free all its associated objects. In this case the TH1Ds objects from Tofsim02
were already destructed which lead to double-freeing of memory and thus to a crash under
certain circumstances (release vs. debug build, libc version, etc..). Fix by a) creating
the TH1Ds on the heap before inserting them to a std::map so that we can use b) SetDirectory(0)
to properly detach the histogram from the ROOT file.

Rubber-stamped by Qi Yan.