Land another harmless patch, which fixes all RICH related GCC + OpenMP linker errors that I was fighting. Doesn't affect ICC builds in any way.
It turns out that you have to start the line with "!$OMP threadprivate..." not with "\s\s\s!$OMP threadprivate..." in include/richrec_lip*.h.

You can use "    !$OMP THREADPRIVATE..." in .F / .f files, but when including the header file containing
the fortran code (like it's the case for include/richrec_lip*.h) the OpenMP pragmas are only recognized when
the line actually starts with "!$OMP...", without leading whitespace. Very subtle issue, was hard to find.