Summary: Fix bug for bad fits with large chi2.
As discussed with Paolo a while ago there was a bug in TrTrack::FitT when the
fit did not work and returned very large chi2. In that case the "retbad" return
code in TrTrack::FitT flipped sign.

Then, the iTrTrackPar that was returned by TrTrack::iTrTrackPar was positive
which is indicative of a good fit - but that was not the case.

The discussed fix was to cap the chi2 at 1e4, but 10e4 were commited
accidentally at the time.

Fix this by capping chi2 at 1e4.