TRD MC: Use realistic gain for each module.
It was found that the gain in the MC of the TRD does not reflect the gain we
have for the majority of the ISS data.

This patch introduces a new TDV called 'TRDMCGains2' which scales the gain in
the MC to the average value observed in ISS (averaged from 01 January 2012 to
the end of pass4) for each module.

Therefore, module-to-module variations are now correctly simulated and the
global scale is also fixed, which mostly matters for the simulation of the
overflow in the tube spectra.

The TRDMCGains2 TDV is time independent: Should the ISS gain every change
significantly we would need to do something about it in the MC as well.

Also fix creation of TrdClusterR in MC: Use TRDMCGains2 to correct relative
module-to-module variations in gain and then use the time average of TRDGains3
to scale the energy deposits to the same scale as in ISS data.

Use TRDMCFFKEY.gain datacard to apply additional gain scaling, for simulation of
tube spectra corresponding to other HV settings for example. This additional
gain is also taken care of when calculating TrdClusterR amplitudes.