Remove references to $(ROOTSYS)/lib and replace them with $(ROOTLIBDIR) Remove a reference to $(ROOTSYS)/include and replace it with $(ROOTINCDIR)
$(ROOTLIBDIR) = root-config --libdir
$(ROOTINCDIR) = root-config --incdir

This is more flexible since when you install root into a directory separate
from the buid/source directory, libraries are installed in $(prefix)/lib/root
and header file are installed in $(prefix)/include/root. This means we can
build gbatch with a root installation which was installed with make install.