Better description of TRD noise.
Several people have noticed that the number of noise hits in the TRD MC is too
large when compared to ISS data. This is because the data reduction was
incorrectly applied to the random noise amplitudes generated in sitrdnoise.

This commit corrects data reduction routine in sitrdnoise:

sigma = idsoft.getsig()   // (typically 1.8 ADC for almost all channels)
sigmaDR = min(5, sigma)   // at most 5 ADC
sigmaDR = max(3, sigmaDR) // at least 3 ADC

cut is on 3*sigmaDR (and used to be on 3*sigma which is wrong.)